Edited test files to include MedicineInventory tests.
