+++ src/test/java/seedu/clinicio/model/ClinicIoTest.java
@@ -8,10 +8,10 @@ import static org.junit.Assert.assertTrue;
 import static seedu.clinicio.logic.commands.CommandTestUtil.VALID_ADDRESS_BOB;
 import static seedu.clinicio.logic.commands.CommandTestUtil.VALID_NAME_ADAM;
 import static seedu.clinicio.logic.commands.CommandTestUtil.VALID_TAG_HUSBAND;
-import static seedu.clinicio.testutil.TypicalMedicines.PARACETAMOL;
-import static seedu.clinicio.testutil.TypicalMedicines.VENTOLIN;
-import static seedu.clinicio.testutil.TypicalMedicines.ORACORT;
-import static seedu.clinicio.testutil.TypicalMedicines.CHLORPHENIRAMINE;
+import static seedu.clinicio.testutil.TypicalPersons.PARACETAMOL;
+import static seedu.clinicio.testutil.TypicalPersons.VENTOLIN;
+import static seedu.clinicio.testutil.TypicalPersons.ORACORT;
+import static seedu.clinicio.testutil.TypicalPersons.CHLORPHENIRAMINE;
 import static seedu.clinicio.testutil.TypicalPersons.ADAM;
 import static seedu.clinicio.testutil.TypicalPersons.ALICE;
 import static seedu.clinicio.testutil.TypicalPersons.AMY_APPT;
